% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{cust_getFpType}
\alias{cust_getFpType}
\title{读取数据信息}
\usage{
cust_getFpType(
  file = "data-raw/专普票区分表-828-V2.xls",
  sheet = "专普",
  lang = "cn"
)
}
\arguments{
\item{file}{文件名}

\item{sheet}{页签}
}
\value{
返回值
}
\description{
读取数据信息
}
\examples{
cust_getFpType()
}
