% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{cust_syncFpType}
\alias{cust_syncFpType}
\title{同步客户的发票数据}
\usage{
cust_syncFpType(
  conn = conn_kjrds(),
  file = "data-raw/专普票区分表-828-V2.xls",
  sheet = "专普"
)
}
\arguments{
\item{conn}{连接}

\item{file}{文件}

\item{sheet}{页答}
}
\value{
无返回值
}
\description{
同步客户的发票数据
}
\examples{
cust_syncFpType()
}
