% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{cust_updateFpType}
\alias{cust_updateFpType}
\title{更新发票数据}
\usage{
cust_updateFpType(conn = conn_kjrds())
}
\arguments{
\item{conn}{连接}
}
\value{
无返回值
}
\description{
更新发票数据
}
\examples{
cust_updateFpType()
}
