% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{cust_uploadFpType}
\alias{cust_uploadFpType}
\title{上传数据}
\usage{
cust_uploadFpType(
  conn = conn_kjrds(),
  file = "data-raw/专普票区分表-828-V2.xls",
  sheet = "专普"
)
}
\arguments{
\item{conn}{连接名}

\item{file}{文件}

\item{sheet}{页答}
}
\value{
返回值
}
\description{
上传数据
}
\examples{
cust_uploadFpType()
}
