% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_customer.R
\name{fp_updateBatch}
\alias{fp_updateBatch}
\title{发票批理更新}
\usage{
fp_updateBatch(conn = tsda::conn_rds("kj613"), data, show_process = FALSE)
}
\arguments{
\item{conn}{连接}

\item{data}{数据}

\item{show_process}{是否显示进度条}
}
\value{
返回值
}
\description{
发票批理更新
}
\examples{
fp_updateBatch()
}
