% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invRpt.R
\name{inv_getMaterial}
\alias{inv_getMaterial}
\title{查询物料安全库存}
\usage{
inv_getMaterial(conn = conn_kjrds(), FCompanyId = 100202)
}
\arguments{
\item{conn}{连接}

\item{FCompanyId}{公司内码}
}
\value{
返回值
}
\description{
查询物料安全库存
}
\examples{
inv_getMaterial()
}
