% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invRpt.R
\name{inv_getRpt}
\alias{inv_getRpt}
\title{获取即时库存报表}
\usage{
inv_getRpt(conn = conn_kjrds(), FCompanyId = 100202)
}
\arguments{
\item{conn}{连接}

\item{FCompanyId}{公司名称}
}
\value{
返回值
}
\description{
获取即时库存报表
}
\examples{
inv_getRpt()
}
