% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invRpt.R
\name{inv_getRpt_cast}
\alias{inv_getRpt_cast}
\title{针对即时库存报表进行处理}
\usage{
inv_getRpt_cast(
  conn = conn_kjrds(),
  FCompanyId = 100202,
  filter_ItemNumber = c("2.01.01.0001", "2.01.01.0008", "3.01.01.0001", "3.01.01.0003",
    "3.01.04.0001", "3.01.04.0002", "3.01.04.0003", "3.01.04.0004", "3.01.04.0005"),
  filter_stockName = c("半成品仓库", "原材料仓库", "采血管车间库",
    "辅料仓库")
)
}
\arguments{
\item{conn}{连接}

\item{FCompanyId}{公司ID}

\item{filter_ItemNumber}{物料代码}

\item{filter_stockName}{物料名称}
}
\value{
返回值
}
\description{
针对即时库存报表进行处理
}
\examples{
inv_getRpt_cast()
}
