% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invRpt.R
\name{inv_paramGetMaterial}
\alias{inv_paramGetMaterial}
\title{返回唯一的物料编码}
\usage{
inv_paramGetMaterial(file = "data-raw/安全库存查询模板.xlsx", sheet = "物料")
}
\arguments{
\item{file}{文件}

\item{sheet}{页答}
}
\value{
返回值
}
\description{
返回唯一的物料编码
}
\examples{
inv_paramGetMaterial()
}
