% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invRpt.R
\name{inv_paramGetStock}
\alias{inv_paramGetStock}
\title{返回唯一的仓库}
\usage{
inv_paramGetStock(file = "data-raw/安全库存查询模板.xlsx", sheet = "仓库")
}
\arguments{
\item{file}{文件}

\item{sheet}{页签}
}
\value{
返回值
}
\description{
返回唯一的仓库
}
\examples{
inv_paramGetStock()
}
