% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_deal_price}
\alias{kjco_deal_price}
\title{材料成本计算单价}
\usage{
kjco_deal_price(conn = conn_kjco(), method = "生产领料单")
}
\arguments{
\item{conn}{连接}

\item{method}{方法}
}
\value{
返回值
}
\description{
材料成本计算单价
}
\examples{
kjco_deal_price()
}
