% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_deal_price_bom}
\alias{kjco_deal_price_bom}
\title{材料出库成本按BOM}
\usage{
kjco_deal_price_bom(conn = conn_kjco())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
材料出库成本按BOM
}
\examples{
kjco_deal_price_bom()
}
