% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_deal_price_pick}
\alias{kjco_deal_price_pick}
\title{材料出库成本按生产领料}
\usage{
kjco_deal_price_pick(conn = conn_kjco())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
材料出库成本按生产领料
}
\examples{
kjco_deal_price_pick()
}
