% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_HumanManufacture_cost_cn}
\alias{kjco_HumanManufacture_cost_cn}
\title{人工及制造费用明细}
\usage{
kjco_HumanManufacture_cost_cn(conn = conn_kjco(), FPeriod = "2020-03")
}
\arguments{
\item{conn}{连接}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
人工及制造费用明细
}
\examples{
kjco_HumanManufacture_cost_cn()
}
