% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_item_lowCode_error}
\alias{kjco_item_lowCode_error}
\title{产品计算步骤异常查询}
\usage{
kjco_item_lowCode_error(conn = conn_kjco())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
产品计算步骤异常查询
}
\examples{
kjco_item_lowCode_error()
}
