% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_materialCost_bom_cn}
\alias{kjco_materialCost_bom_cn}
\title{材料成本生产领料口径中文标题}
\usage{
kjco_materialCost_bom_cn(conn = conn_kjco(), FPeriod = "2020-03")
}
\arguments{
\item{conn}{连接}

\item{FPeriod}{年月}
}
\value{
返回值
}
\description{
材料成本生产领料口径中文标题
}
\examples{
kjco_materialCost_bom_cn()
}
