% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_materialCost_pick}
\alias{kjco_materialCost_pick}
\title{材料成本按生产领料单口径}
\usage{
kjco_materialCost_pick(conn = conn_kjco(), FPeriod = "2020-03")
}
\arguments{
\item{conn}{连接}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
材料成本按生产领料单口径
}
\examples{
kjco_materialCost_pick()
}
