% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_materialCost_ppbom_cn}
\alias{kjco_materialCost_ppbom_cn}
\title{材料成本按生产用料清单}
\usage{
kjco_materialCost_ppbom_cn(conn = conn_kjco(), FPeriod = "2020-03")
}
\arguments{
\item{conn}{连接}

\item{FPeriod}{年月}
}
\value{
返回值
}
\description{
材料成本按生产用料清单
}
\examples{
kjco_materialCost_ppbom_cn()
}
