% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_prdCost_3Compare}
\alias{kjco_prdCost_3Compare}
\title{生产成本数据}
\usage{
kjco_prdCost_3Compare(conn = conn_kjco())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
生产成本数据
}
\examples{
kjco_prdCost_3Compare()
}
