% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjco.R
\name{kjco_prdCost_3Compare_cn}
\alias{kjco_prdCost_3Compare_cn}
\title{计算成本差异中文版}
\usage{
kjco_prdCost_3Compare_cn(conn = conn_kjco())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
计算成本差异中文版
}
\examples{
kjco_prdCost_3Compare_cn()
}
