% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{mtrl_bakFpName}
\alias{mtrl_bakFpName}
\title{备份物料上的开票物料名称及规格型号}
\usage{
mtrl_bakFpName(conn = conn_kjrds())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
备份物料上的开票物料名称及规格型号
}
\examples{
mtrl_bakFpName()
}
