% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{mtrl_getFpName}
\alias{mtrl_getFpName}
\title{读取物料的开票名称及开票规格型号}
\usage{
mtrl_getFpName(
  file = "data-raw/2019年新K3开票名称-V3.XLS",
  sheet = "开票用最新",
  lang = "cn"
)
}
\arguments{
\item{file}{文件}

\item{sheet}{页签}

\item{lang}{语言}
}
\value{
返回值
}
\description{
读取物料的开票名称及开票规格型号
}
\examples{
mtrl_getFpName()
}
