% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{mtrl_updateFpName}
\alias{mtrl_updateFpName}
\title{更新物料发票信息}
\usage{
mtrl_updateFpName(conn = conn_kjrds())
}
\arguments{
\item{conn}{连接}
}
\value{
无返回值
}
\description{
更新物料发票信息
}
\examples{
mtrl_updateFpName()
}
