% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{mtrl_uploadFpName}
\alias{mtrl_uploadFpName}
\title{物料的开票名称上传}
\usage{
mtrl_uploadFpName(
  conn = conn_kjrds(),
  file = "data-raw/2019年新K3开票名称-V3.XLS",
  sheet = "开票用最新"
)
}
\arguments{
\item{conn}{连接}

\item{file}{文件}

\item{sheet}{页签}
}
\value{
返回值
}
\description{
物料的开票名称上传
}
\examples{
mtrl_uploadFpName()
}
