% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_customer.R
\name{fp_update}
\alias{fp_update}
\title{更新发票信息}
\usage{
fp_update(
  conn = tsda::conn_rds("kj613"),
  custName = "北京健辰医疗器械有限公司",
  bank = "bankName",
  acctNumber = "acctNumber",
  phone = "phone",
  taxNo = "taxNo",
  addr = "addr"
)
}
\arguments{
\item{conn}{连接}

\item{phone}{电话}

\item{taxNo}{税号}

\item{fname}{名称}

\item{bankName}{银行名词}

\item{bankAccount}{账号}

\item{fpAddr}{发票地址}
}
\value{
返回值
}
\description{
更新发票信息
}
\examples{
fp_update()
}
